Adding unit tests
