+++ package.json
@@ -47,6 +47,7 @@
     "@types/react": "^16.7.13",
     "@types/react-dom": "^16.0.9",
     "@types/react-resize-detector": "^4.0.1",
+    "@types/react-test-renderer": "^16.8.1",
     "@types/webdriverio": "^4.13.0",
     "mendixmodelsdk": "4.19.0",
     "react": "^16.6.3",
