+++ src/components/Error.tsx
@@ -4,7 +4,7 @@ interface PlayerErrorProps {
     preview?: boolean;
 }
 
-export const PlayerError: FunctionComponent<PlayerErrorProps> = (props: PlayerErrorProps): JSX.Element => (
+export const Error: FunctionComponent<PlayerErrorProps> = (props: PlayerErrorProps): JSX.Element => (
     <div className="widget-video-player-error-container">
         <div className="video-error-label">
             {props.preview ? "The actual video rendering will be based on the attribute content." : "The video failed to load :("}
