+++ src/components/__tests__/Alert.spec.tsx
@@ -1,19 +1,19 @@
-import { shallow } from "enzyme";
 import { createElement } from "react";
 
 import { Alert } from "../Alert";
+import { create } from "react-test-renderer";
 
 describe("Alert", () => {
     it("renders the structure when an alert message is specified", () => {
         const message = "This is an error";
-        const alert = shallow(<Alert bootstrapStyle="danger" className="widget-badge-alert" message={message} />);
+        const alert = create(<Alert bootstrapStyle="danger" className="widget-badge-alert" message={message} />).toJSON();
 
-        expect(alert.equals(<div className="alert alert-danger widget-badge-alert">{message}</div>)).toEqual(true);
+        expect(alert).toMatchSnapshot();
     });
 
     it("renders no structure when the alert message is not specified", () => {
-        const alert = shallow(<Alert />);
+        const alert = create(<Alert />).toJSON();
 
-        expect(alert.isEmptyRender()).toEqual(true);
+        expect(alert).toBeNull();
     });
 });
