+++ src/components/__tests__/Error.spec.tsx
@@ -1,0 +1,11 @@
+import { createElement } from "react";
+
+import { PlayerError } from "../PlayerError";
+import { create } from "react-test-renderer";
+
+describe("Player Error", () => {
+    it("should renders correctly", () => {
+        const player = create(<PlayerError />).toJSON();
+        expect(player).toMatchSnapshot();
+    });
+});
