+++ src/components/__tests__/Youtube.spec.tsx
@@ -1,93 +1,53 @@
 import { createElement } from "react";
-import { shallow } from "enzyme";
 
+import { Youtube, YoutubeProps } from "../Youtube";
+import { create } from "react-test-renderer";
 import ReactResizeDetector from "react-resize-detector";
-import { YoutubePlayer } from "../YoutubePlayer";
 
 describe("YoutubePlayer Player", () => {
-    it("Renders the structure of iframe tags and check classes", () => {
-        const player = shallow(
-            <YoutubePlayer url="http://youtube.com/watch?v=123456" autoPlay={false} muted={false} loop={false} showControls={false} aspectRatio={false} />,
-        );
+    const defaultProps = {
+        url: "http://youtube.com/watch?v=123456",
+        autoPlay: false,
+        muted: false,
+        loop: false,
+        showControls: false,
+        aspectRatio: false,
+    };
 
-        expect(player).toHaveClass("widget-video-player-iframe");
-        expect(player).toBeDefined();
-    });
-    it("Renders the structure of iframe tags and check the structure", () => {
-        const player = shallow(
-            <YoutubePlayer url="http://youtu.be/123456" autoPlay={false} muted={false} loop={false} showControls={false} aspectRatio={false} />,
-        );
+    const defaulPlayer = (props: YoutubeProps) => <Youtube {...props} />;
 
-        expect(
-            player.equals(
-                <iframe
-                    className="widget-video-player-iframe"
-                    frameBorder="0"
-                    allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-                    allowFullScreen={true}
-                    src="https://www.youtube.com/embed/123456?modestbranding=1&rel=0&autoplay=0&controls=0&muted=0&loop=0"
-                >
-                    <ReactResizeDetector handleWidth handleHeight onResize={jasmine.any(Function)} refreshMode="debounce" refreshRate={100} />
-                </iframe>,
-            ),
-        ).toEqual(true);
+    it("should renders correctly", () => {
+        const player = create(defaulPlayer(defaultProps)).toJSON();
+
+        expect(player).toMatchSnapshot();
     });
 
-    it("Renders the structure of iframe tags and check the structure with parameters true", () => {
-        const player = shallow(
-            <YoutubePlayer url="http://youtube.com/watch?v=123456" autoPlay={true} muted={true} loop={true} showControls={true} aspectRatio={true} />,
-        );
+    it("should renders correctly with autoplay", () => {
+        const player = create(defaulPlayer({ ...defaultProps, autoPlay: true })).toJSON();
 
-        expect(
-            player.equals(
-                <iframe
-                    className="widget-video-player-iframe"
-                    frameBorder="0"
-                    allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-                    allowFullScreen={true}
-                    src="https://www.youtube.com/embed/123456?modestbranding=1&rel=0&autoplay=1&controls=1&muted=1&loop=1"
-                >
-                    <ReactResizeDetector handleWidth handleHeight onResize={jasmine.any(Function)} refreshMode="debounce" refreshRate={100} />
-                </iframe>,
-            ),
-        ).toEqual(true);
+        expect(player).toMatchSnapshot();
     });
 
-    it("Renders the structure of iframe tags and check the structure with embed URL", () => {
-        const player = shallow(
-            <YoutubePlayer url="http://youtube.com/embed/123456" autoPlay={true} muted={true} loop={true} showControls={true} aspectRatio={true} />,
-        );
+    it("should renders correctly with muted", () => {
+        const player = create(defaulPlayer({ ...defaultProps, muted: true })).toJSON();
 
-        expect(
-            player.equals(
-                <iframe
-                    className="widget-video-player-iframe"
-                    frameBorder="0"
-                    allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-                    allowFullScreen={true}
-                    src="http://youtube.com/embed/123456?modestbranding=1&rel=0&autoplay=1&controls=1&muted=1&loop=1"
-                >
-                    <ReactResizeDetector handleWidth handleHeight onResize={jasmine.any(Function)} refreshMode="debounce" refreshRate={100} />
-                </iframe>,
-            ),
-        ).toEqual(true);
+        expect(player).toMatchSnapshot();
     });
 
-    it("Renders the structure of iframe tags and check the structure with empty url", () => {
-        const player = shallow(<YoutubePlayer url="" autoPlay={true} muted={true} loop={true} showControls={true} aspectRatio={true} />);
+    it("should renders correctly with controls", () => {
+        const player = create(defaulPlayer({ ...defaultProps, showControls: true })).toJSON();
 
-        expect(
-            player.equals(
-                <iframe
-                    className="widget-video-player-iframe"
-                    frameBorder="0"
-                    allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-                    allowFullScreen={true}
-                    src=""
-                >
-                    <ReactResizeDetector handleWidth handleHeight onResize={jasmine.any(Function)} refreshMode="debounce" refreshRate={100} />
-                </iframe>,
-            ),
-        ).toEqual(true);
+        expect(player).toMatchSnapshot();
+    });
+
+    it("should renders correctly with aspectRatio", () => {
+        const player = create(defaulPlayer({ ...defaultProps, aspectRatio: true }));
+        window.dispatchEvent(new Event("resize"));
+        const instance = player.root;
+        const sizeDetector = instance.findByType(ReactResizeDetector);
+
+        expect(player.toJSON).toMatchSnapshot();
+        expect(sizeDetector).not.toBeNull();
+        expect(sizeDetector.props).toHaveProperty("onResize");
     });
 });
